test(access): fix quality gaps in access/policy test suites
Remove experimental spike tests and strengthen assertions.
